# back/backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.ChatMessage:
    properties:
      applicationID:
        format: int64
        type: integer
      createdAt:
        type: string
      id:
        format: int64
        type: integer
      message:
        type: string
      senderID:
        format: int64
        type: integer
    type: object
  domain.Vacancy:
    properties:
      createdAt:
        type: string
      description:
        type: string
      employerID:
        format: int64
        type: integer
      id:
        format: int64
        type: integer
      isActive:
        type: boolean
      salaryFrom:
        format: int32
        type: integer
      salaryTo:
        format: int32
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      workFormat:
        $ref: '#/definitions/domain.WorkFormat'
    type: object
  domain.WorkFormat:
    enum:
    - remote
    - hybrid
    - office
    - project
    type: string
    x-enum-varnames:
    - WorkRemote
    - WorkHybrid
    - WorkOffice
    - WorkProject
  handler.createEmployerRequest:
    properties:
      company_name:
        type: string
      description:
        type: string
      inn:
        type: string
      website:
        type: string
    required:
    - company_name
    - inn
    type: object
...
  handler.createVacancyRequest:
    properties:
      description:
        type: string
      salary_from:
        type: integer
      salary_to:
        type: integer
      title:
        type: string
      work_format:
        type: string
    required:
    - description
    - salary_from
    - salary_to
    - title
    - work_format
    type: object
...
  handler.sendMessageRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  vacancy.MatchResponse:
    properties:
      missingSkills:
        items:
          $ref: '#/definitions/vacancy.SkillResponse'
        type: array
      recommendation:
        type: string
      score:
        type: integer
    type: object
  vacancy.SkillResponse:
    properties:
      id:
        type: integer
      isRequired:
        type: boolean
      name:
        type: string
      requiredLevel:
        type: integer
    type: object
  vacancy.VacancyDetailResponse:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      match:
        $ref: '#/definitions/vacancy.MatchResponse'
      salaryFrom:
        type: integer
      salaryTo:
        type: integer
      skills:
        items:
          $ref: '#/definitions/vacancy.SkillResponse'
        type: array
      status:
        type: string
      title:
        type: string
      workFormat:
        type: string
    type: object
  vacancy.VacancyListItemResponse:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      salaryFrom:
        type: integer
      salaryTo:
        type: integer
      status:
        type: string
      title:
        type: string
      workFormat:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Skill-first hiring platform backend.
  title: Hunter Platform API
  version: "1.0"
paths:
  /applications/{id}/chat:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ChatMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - chat
    post:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.sendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send chat message
      tags:
      - chat
...
  /employer/profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Employer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createEmployerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create employer profile
      tags:
      - employer
  /employer/vacancies:
    post:
      consumes:
      - application/json
      parameters:
      - description: Vacancy data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createVacancyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Vacancy'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create vacancy
      tags:
      - vacancy
...
  /vacancies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vacancy.VacancyListItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List public vacancies
      tags:
      - vacancy
  /vacancies/{id}:
    get:
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vacancy.VacancyDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get vacancy details
      tags:
      - vacancy
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"